+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Fixed
 
 Changed
 - Use google-cloud-php's CacheSessionPool since the [concerned bug](https://github.com/googleapis/google-cloud-php/issues/5567) has been fixed in [v1.53](https://github.com/googleapis/google-cloud-php-spanner/releases/tag/v1.58.2). (#90)
+- Separate session pool and authentication per connection so transaction works properly.
 
 # v5.0.0